Clean up action-classifier and Gym-eval pipeline, remove dead code
